Use System.in rather than FileDescriptor.in
In sbt, we want to use System.in rather than FileDescriptor.in here
because the scala console calls this api and we need the console to read
from System.in rather than FileDescriptor.in. Otherwise the console can
fight with sbt's inputstream and bytes can be lost. It is necessary to
use wrapper classes for the streams because jline will perform its own
wrapping of the default input and output streams (System.in and
System.out). sbt sets System.in and System.out to streams that have
already been wrapped by jline and the double wrapping breaks the i/o on
windows.