Upgrade jansi to latest version
The latest version of jansi that jline 3 now uses breaks binary
compatibility relative to the jline 2 jansi version. This commit both
bumps the jansi version and fixes the broken binary compatibility. The
advantage to this is that we should hopefully be able to upgrade jline3
(and transitively jansi) in the future without worrying about breaking
the 2.12 console and the thin client (which graalvm has problems
compiling with the missing apis).